Refactor trace source for Tracer
