+++ tests/test_tracer.py
@@ -4,6 +4,7 @@ import tracer
 
 import logging
 l = logging.getLogger("tracer.tests.test_tracer")
+logging.getLogger('tracer').setLevel('DEBUG')
 
 bin_location = str(os.path.join(os.path.dirname(os.path.realpath(__file__)), '../../binaries'))
 pov_location = str(os.path.join(os.path.dirname(os.path.realpath(__file__)), "povs"))
